Add helpful error messaging to table name and calls
Adds more helpful error messages if the table provided is invalid or
if it does not exist. If the table does not exist, the user could get
a generic ets error or failed GenServer call error but it would not
be apparent it is a result of bad user table id input without the user
knowing the internals of PropertyTable.