Minor optimization to dispatch
Replace use of `&&` with `and`. The former is a `case`..`do` with a
match (see
https://github.com/elixir-lang/elixir/blob/v1.13.4/lib/elixir/lib/kernel.ex#L3925-L3931)
and the latter is one call to `:erlang.andalso`. See
https://github.com/elixir-lang/elixir/blob/v1.13.4/lib/elixir/lib/kernel.ex#L1824.

The left side is the only side that needs to be boolean, so this works.
It results in a small performance improvement since it's in the critical
dispatch loop.