Standardize all GenServer naming on the table name
This moves the public API for starting a PropertyTable completely to
`property_table.ex`. It also changes the names in the process tree so
that users are adding to their namespace rather than PropertyTable's. It
looks like this:

```
<table_name>
<table_name>.Subscribers
<table_name>.Table
```

The `<table_name>` process is the supervisor. If the user calls
`GenServer.stop` on it, everything goes down like you'd expect. This is
different from before. Previously the PropertyTable supervisor was
unnamed.

All parameter validation is in `PropertyTable.start_link/1`. The code in
`PropertyTable.Supervisor` is minimal now and it's been removed from the
public API.