Use :ets.foldl to avoid long temporary lists
Previously, `:ets.match` was being used to return a list of possible
matches for `PropertyTable.get_all/3` and `PropertyTable.match/2`. These
were then filtered and adjusted for return to the user. In both cases,
`:ets` was iterating over every property. This doesn't change that, but
now there's one pass that does everything and no temporaries.