Extract matching logic to enable alternative keys
This is the first step in extracting the matching logic to its own
module. The goal is to make PropertyTable useful if you'd like to just
plain strings or maps for keys or if you'd like to have a custom
pattern matcher.

The one big change for hierarchical string lists is to make "prefix"
matching the default for everything. If you'd like to force the number
of fields in the property use the `:"$"` atom to signify that there
should be no more elements in the property.