Remove special handling for nil
Storing `nil` in a table can be useful for maintaining timestamps on
properties that come and go. This also matches the semantics of other
Elixir KV stores with regards to `nil`, so hopefully it better matches
people's expectations when using `PropertyTable`.

This feature originated since events generated from setting a property
the first time, clearing it, and setting it to `nil` were ambigous.
Setting properties the first time stopped being ambigous when timestamps
were added. If it becomes to disambiguate clear the property from
setting it to `nil`, then that could be fixed by adding a "reason" field
to the events.