Raise on argument error to put
`PropertyTable.put/3` used to return an error if the property didn't
comform to what the `Matcher` implementation expected. This turned out
to be really annoying with Dialyzer in that you had to check the return
value every single time you used `put`. Since messing up a property
feels like it a programming error most of the time, using an exception
here seems like a better choice.