Rename Table module to Updater and move reader code
This change makes it more obvious what the role of Table (aka Updater)
is: anything that wants to change the contents of a property table
must end up going through this module. All writes are in here.

This change also moves almost all reads of the table out. This removes
the need for many defdelegates with their repeated docs and specs.

This doesn't add or remove any functionality.