Minimize disturbance to persisted files while saving
This introduces a temporary file that's constructed while saving the ETS
table to disk when persisting. Only after the file is created are the
main and backup files renamed and deleted. The theory is that saving the
ETS file is by far the slowest part of the process. Renaming and
deleting files is expected to be fast. Therefore, if power is lost or
anything else weird happens, it most likely affects the temporary file
which will never be read anyway.