Set event previous timestamps as best as possible
Previously when events were sent on a property for the first time, the
`previous_timestamp` was `nil`. This was a typespec violation, but it
also made it hard to know how long the property was in the unknown
state. Depending on the use, an application could infer a default value
on the property. It could also assign an "unknown" value. Either way, it
could be useful to know how long it was in this state and `nil` didn't
help.

A side effect of this change is that all values set explicitly or
implicitly at startup will have the same timestamp. This could come in
handy.